back/pyproject.toml
[tool.poetry]
name = "fast-api-sample"
version = "0.1.0"
description = ""
authors = []

readme = "README.md"
packages = [
    { include = "src", from = "." },
    { include = "app", from = "." }
]

[tool.poetry.dependencies]
python = "~3.10"
fire = "^0.5.0"
fastapi = "^0.110.2"
mypy = "^1.11.1"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
uvicorn = "^0.30.6"
openai = "^1.43.0"
pydantic-settings = "^2.4.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.2"
mypy = "^1.11.1"
pytest = "^7.2.0"
google-auth-stubs = "^0.2.0"
google-api-python-client-stubs = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

